requirements.txt
numpy==1.16.2
tensorflow==1.14.0
opencv_python==4.1.0.25
pandas==0.24.2
gym==0.14.0
Keras==2.2.4
matplotlib==3.0.3
Pillow==6.1.0
